// frontend/src/shared/api/model/updatePlaceRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Tarikma API
 * Backend API for Tarik.ma
 * OpenAPI spec version: v1
 */

export interface UpdatePlaceRequest {
  /**
   * @minLength 0
   * @maxLength 300
   */
  name?: string;
  description?: string;
  /**
   * @minLength 0
   * @maxLength 500
   */
  image?: string;
  /**
   * @minLength 0
   * @maxLength 500
   */
  address?: string;
  /**
   * @minLength 0
   * @maxLength 100
   */
  category?: string;
  latitude?: number;
  longitude?: number;
  /**
   * @minLength 0
   * @maxLength 255
   */
  mapPlaceId?: string;
}